feat: Exception handling

+++ src/main/java/io/github/Ital023/dscommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package io.github.Ital023.dscommerce.services;
 import io.github.Ital023.dscommerce.dto.ProductDTO;
 import io.github.Ital023.dscommerce.entities.Product;
 import io.github.Ital023.dscommerce.repositories.ProductRepository;
+import io.github.Ital023.dscommerce.services.exceptions.DatabaseException;
 import io.github.Ital023.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
@@ -50,19 +54,32 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(ProductDTO productDTO, Long id) {
+        try {
+            Product entity = productRepository.getReferenceById(id);
 
-        Product entity = productRepository.getReferenceById(id);
+            copyDtoToEntity(productDTO, entity);
 
-        copyDtoToEntity(productDTO, entity);
+            entity = productRepository.save(entity);
 
-        entity = productRepository.save(entity);
+            return new ProductDTO(entity);
+        }catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
 
-        return new ProductDTO(entity);
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        productRepository.deleteById(id);
+        if(!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+
+        try{
+            productRepository.deleteById(id);
+        }catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha na integridade referencial");
+        }
+
     }
 
     private void copyDtoToEntity(ProductDTO productDTO, Product entity) {
